requirements.txt
Flask>=3.0.3
google-cloud-storage>=2.18.2
google-generativeai>=0.8.3
